requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==4.2.25
django-cors-headers==4.3.1
djangorestframework==3.15.2
idna==3.11
python-dotenv==1.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
requests==2.31.0
urllib3==1.26.18
Pillow==9.5.0
reportlab==3.6.13

